WHO IS ME?
Cobweb Coin is a deflationary token, that is, in every transaction, be it a purchase or a sale, a part is sent to the Burn Address, thus making the token scarce in the long run and enabling it to become more valuable. Each time a buy or sell is made a 10% fee is charged, with 5% returns to holders and 5% percent goes to the Liquidity Pool, this mechanism helps to increase token appreciation and rewards all hand holders diamond.